+++ pages/blogzoom.js
@@ -2,16 +2,19 @@ import React from 'react'
 import Link from 'next/link'
 import * as contentful from 'contentful'
 import { config } from '../src/config'
+import BlogZoomTemplate from '../src/components/templates/blogzoom';
 
 const PostItem = props => {
-  const post = props.item.fields
+  // const post = props.item.fields
+  console.log(props)
 
   return (
     <div>
+    <BlogZoomTemplate />
       <h1>Post detail</h1>
       <div>
-        <h2>{post.title}</h2>
-        <p>{post.body}</p>
+        {/* <h2>{post.title}</h2>
+        <p>{post.body}</p> */}
       </div>
       <Link prefetch href="/posts"><a>Post一覧へ</a></Link>
     </div>
